Refactor webcamMode using inherritance
